fixing responsiveness Header e NavBar
