--- src/components/Navbar/styles.ts
@@ -1,104 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-`;
-
-export const SecondContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background-color: ${ props => props.theme.title === 'light' ? props.theme.colors.blue.quaternary : props.theme.colors.primary };
-  width: 100%;
-  /* height: 30%;  PERGUNTAR SE PODEMOS DEIXAR COMO PIXEL :  pode sim, só ver a responsividade*/ 
-  height: 130px;
-`;
-
-export const Logo = styled.img`
-  margin-top: 12px;
-  width: 100px;
-  height: 50px;
-  cursor: pointer;
-  object-fit: contain;
-`;
-
-export const ButtonNavbar = styled.h3`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-right: 4cqb;
-  margin-top: 20px;
-  border: none;
-  font-size: 20px;
-  cursor: pointer;
-  color: ${props => props.theme.fontsSizes.colors.white};
-  font-weight: bold;
-  width: 180px;
-  height: 40px;
-  transition: 0.4s;
-  border-radius: 5px;
-  display: block;
-  &:hover {
-    transform: scale(1.1);
-  }
-
-  @media (max-width: 768px) {
-    display: none;
-  }
-`;
-
-export const ButtonIcon = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-right: 20px;
-  margin-top: 20px;
-  background-color: transparent;
-  border: none;
-  font-size: 20px;
-  cursor: pointer;
-  font-weight: bold;
-  width: 40px;
-  height: 40px;
-  transition: 0.4s;
-  border-radius: 5px;
-  display: none;
-  &:hover {
-    transform: scale(1.1);
-  }
-
-  @media (max-width: 768px) {
-    display: block;
-  }
-
-`;
-
-export const ThemeIcon = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-right: 20px;
-  margin-top: 20px;
-  background-color: transparent;
-  border: none;
-  font-size: 20px;
-  cursor: pointer;
-  font-weight: bold;
-  width: 40px;
-  height: 40px;
-  transition: 0.4s;
-  border-radius: 5px;
-  display: block;
-  &:hover {
-    transform: scale(1.1);
-  }
-
-  @media (max-width: 768px) {
-    display: none;
-  }
-
-`;
-
-export const Linha = styled.div`
-  height: 4px;
-  background-color: ${ props => props.theme.colors.blue.quaternary };
-`;
